+++ clients/js/.eslintrc.cjs
@@ -20,5 +20,6 @@ module.exports = {
   rules: {
     '@typescript-eslint/no-explicit-any': 'off',
     '@typescript-eslint/no-non-null-assertion': 'off',
+    eqeqeq: ['warn', 'always'],
   },
 };